% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.dist.R
\name{ds.dist}
\alias{ds.dist}
\title{Computes the distance matrix of a given data set}
\usage{
ds.dist(
  df.name = NULL,
  method = "euclidean",
  newobj = NULL,
  datasources = NULL
)
}
\arguments{
\item{df.name}{is a string character of the data set (either a data.frame or a matrix)}

\item{method}{specifies the method for the distance matrix calculation and can be either 'euclidean', 'maximum', 'manhattan', 'canberra', 'binary' or 'minkowski'}

\item{newobj}{is the name of the new object which is created with this function}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login}
}
\value{
the object specified by the \code{newobj} argument of \code{ds.dist} or default name \code{dist.newobj}
}
\description{
This function is similar to the native R function from stats
}
\details{
The function calls the server-side function \code{distDS} that computes the
distance matrix of a data set (type data.frame or matrix) and assigns the new object to the server-side, which will be of class 'dist'.
The new object is named by the user using the \code{newobj} argument, otherwise it is named \code{dist.newobj} by default.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
