% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.hclust.R
\name{ds.hclust}
\alias{ds.hclust}
\title{Conducts hierarchical clustering of a given dissimilarity structure}
\usage{
ds.hclust(diss = NULL, method = "ward.D2", newobj = NULL, datasources = NULL)
}
\arguments{
\item{diss}{is a string character of the dissimilarity structure}

\item{method}{specifies the method for the calculation of the hierarchical clustering and can be either 'ward.D', 'ward.D2', 'single', 
'complete', 'average', 'mcquitty', 'median' or 'centroid'}

\item{newobj}{is the name of the new object which is created with this function}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login}
}
\value{
the object specified by the \code{newobj} argument of \code{ds.hclust} or default name \code{hclust.newobj}
}
\description{
This function is similar to the native R function 'hclust' from stats
}
\details{
The function calls the server-side function \code{hclustDS} that computes the
hierarchical clustering of a dissimilarity structure (type 'dist'). The dissimilarity structure can be computed via the ds.dist function.
The function creates a new object on the server-side, which is of class 'hclust'.
The new object is named by the user using the \code{newobj} argument, otherwise it is named \code{hclust.newobj} by default.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
