% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.kmeans.R
\name{ds.kmeans}
\alias{ds.kmeans}
\title{Conducts k-means clustering of a given data set}
\usage{
ds.kmeans(
  df.name = NULL,
  clusters = NULL,
  iter.max = 10L,
  nstart = 1L,
  algorithm = "Hartigan-Wong",
  trace = FALSE,
  newobj = NULL,
  datasources = NULL
)
}
\arguments{
\item{df.name}{is a string character of the data set}

\item{clusters}{specifies the number of clusters for the computation}

\item{iter.max}{specifies the max. number of iterations allowed}

\item{nstart}{relates to the number of random sets if clusters is a number and not a set of initial cluster centers}

\item{algorithm}{refers to the algorithm of calculating the kmeans and can be either 'Hartigan-Wong', 'Lloyd', 'Forgy' or 'MacQueen'}

\item{newobj}{is the name of the new object which is created with this function}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login}
}
\value{
the object specified by the \code{newobj} argument of \code{ds.kmeans} or default name \code{kmeans.newobj}
}
\description{
This function is similar to the native R function 'kmeans' from stats
}
\details{
The function calls the server-side function \code{kmeansDS} that computes the
k-means clustering of a data set (type 'data.frame' or 'matrix'). 
The function creates a new object on the server-side, which is of class 'kmeans'.
The new object is named by the user using the \code{newobj} argument, otherwise it is named \code{kmeans.newobj} by default.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
