% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.varSelLcm_Test.R
\name{ds.varSelLcm_Test}
\alias{ds.varSelLcm_Test}
\title{Conducts k-means clustering of a given data set}
\usage{
ds.varSelLcm_Test(
  df = NULL,
  num.clust = NULL,
  vbleSelec = TRUE,
  crit.varsel = "BIC",
  initModel = 50,
  nbcores = 1,
  nbSmall = 250,
  iterSmall = 20,
  nbKeep = 50,
  iterKeep = 1000,
  tolKeep = 1e-07,
  num.iterations = 5,
  newobj = NULL,
  datasources = NULL
)
}
\arguments{
\item{df}{is a string character of the data set}

\item{num.clust}{specifies the number of clusters for the computation}

\item{vbleSelec}{specifies the max. number of iterations allowed}

\item{crit.varsel}{relates to the number of random sets if clusters is a number and not a set of initial cluster centers}

\item{initModel}{refers to the algorithm of calculating the kmeans and can be either 'Hartigan-Wong', 'Lloyd', 'Forgy' or 'MacQueen'}

\item{nbcores}{is the name of the new object which is created with this function}

\item{nbSmall}{is a logical or integer specifying whether tracing information on the progress of the algorithm is procuded for the Hartigan-Wong algorithm}

\item{iterSmall}{is a logical or integer specifying whether tracing information on the progress of the algorithm is procuded for the Hartigan-Wong algorithm}

\item{nbKeep}{is a logical or integer specifying whether tracing information on the progress of the algorithm is procuded for the Hartigan-Wong algorithm}

\item{iterKeep}{is a logical or integer specifying whether tracing information on the progress of the algorithm is procuded for the Hartigan-Wong algorithm}

\item{tolKeep}{represents the number at which point two successive models are defined to be converged; default is 1e-7}

\item{num.iterations}{the number of iterations for finding SLMA clusters in each respective datasource}

\item{newobj}{is a logical or integer specifying whether tracing information on the progress of the algorithm is procuded for the Hartigan-Wong algorithm}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login}

\item{discrim}{is a logical or integer specifying whether tracing information on the progress of the algorithm is procuded for the Hartigan-Wong algorithm}
}
\value{
the object specified by the \code{newobj} argument of \code{ds.kmeans} or default name \code{kmeans.newobj}
}
\description{
This function is similar to the native R function 'kmeans' from stats
}
\details{
The function calls the server-side function \code{kmeansDS} that computes the
k-means clustering of a data set (type 'data.frame' or 'matrix'). 
The function creates a new object on the server-side, which is of class 'kmeans'.
The new object is named by the user using the \code{newobj} argument, otherwise it is named \code{kmeans.newobj} by default.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
