% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.clusterPlot.R
\name{ds.clusterPlot}
\alias{ds.clusterPlot}
\title{Draws a basic plot for clustering methods}
\usage{
ds.clusterPlot(
  tree = NULL,
  k = NULL,
  h = NULL,
  k_colors = NULL,
  palette = NULL,
  show_labels = TRUE,
  color_labels_by_k = FALSE,
  datasources = NULL
)
}
\arguments{
\item{tree}{is a string character of the data set}

\item{k}{specifies the number of clusters in which the tree should be cut}

\item{h}{specifies the height of a tree at which the tree should be cut}

\item{k_colors}{is a vector containing colors to be used for groups}

\item{palette}{a vector containing colors to be used for groups}

\item{show_labels}{will always be set to false on the server-side for disclosure reasons}

\item{color_labels_by_k}{is a logical value which colors the branches by group when k is not NULL}

\item{datasources}{DSCOnnections}
}
\value{
the object specified by the \code{newobj} argument of \code{ds.kmeans} or default name \code{kmeans.newobj}
}
\description{
This function produces a plot for hclust clustering data
}
\details{
The function calls the server-side function \code{kmeansDS} that computes the
k-means clustering of a data set (type 'data.frame' or 'matrix'). 
The function creates a new object on the server-side, which is of class 'kmeans'.
The new object is named by the user using the \code{newobj} argument, otherwise it is named \code{kmeans.newobj} by default.

The function computes the dendrogram without any labels to prevent any disclosures.
The new object is named by the user using the \code{newobj} argument, otherwise it is named \code{kmeans.newobj} by default.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
