% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.nbclust.R
\name{ds.nbclust}
\alias{ds.nbclust}
\title{Determines optimal number of clusters for dataset}
\usage{
ds.nbclust(
  df.name = NULL,
  diss = NULL,
  distance = "euclidean",
  min.nc = 2,
  max.nc = 15,
  method = NULL,
  index = "all",
  alphaBeale = 0.1,
  datasources = NULL
)
}
\arguments{
\item{df.name}{is a string character of the data set and can be either a matrix or data frame}

\item{diss}{is a dissimilarity structure which will be calculated according to the distance method}

\item{distance}{specifies the method for the distance matrix calculation and can be either 'euclidean', 'maximum', 'manhattan', 'canberra', 'binary' or 'minkowski'}

\item{min.nc}{specifies the minimum number of clusters}

\item{max.nc}{specifies the maximum number of clusters}

\item{method}{describes the clustering method and can be either "ward.D2", "single", "complete", "average", "mcquitty", "median", "centroid", "kmeans" or "ward.D"}

\item{index}{describes the clustering index and can be either "kl", "ch", "hartigan", "ccc", "scott", "marriot", "trcovw", "tracew", "friedman", "rubin", "cindex", "db", "silhouette", "duda", "pseudot2", "beale", "ratkowsky", "ball", "ptbiserial", "gap", "frey", "mcclain", "gamma", "gplus", "tau", "dunn", "hubert", "sdindex", "dindex", "sdbw", "all" or "alllong"}

\item{alphaBeale}{value for "beale" clustering index}
}
\value{
a summary suggesting the optimal number of clusters
}
\description{
This function is similar to the R function 'NBClust' from the NBClust package
}
\details{
The function uses partitioning methods to find optimal numbers of clusters for a given dataset.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
